requirements.txt
# Environment setup instructions for MacOS with Conda
# conda create -n hubert_tokenizer python=3.11
# conda activate hubert_tokenizer

# Core
numpy==1.26.4
torch==2.1.2
torchaudio==2.1.2

# Audio processing
soundfile==0.13.1
scipy==1.16.3
sounddevice==0.4.7
librosa==0.11.0
soxr==1.0.0
audioread==3.1.0
pooch==1.8.2

# Transformers / LLM token projection
transformers==4.34.0
huggingface_hub==0.17.1
tokenizers==0.14.1
einops==0.8.1

# Utilities
requests==2.32.5
tqdm==4.67.1
platformdirs==4.4.0
msgpack==1.1.2
lazy_loader==0.4
decorator==5.2.1
certifi==2025.11.12
charset_normalizer==3.4.4
idna==3.11
faiss-cpu==1.9.0
